fix:別のempathyボタンを押下できる

+++ app/controllers/empathies_controller.rb
@@ -3,14 +3,15 @@ class EmpathiesController < ApplicationController
   def create
     post = Post.find(params[:post_id])
     empathy_value = Empathy.empathies[params[:empathy]] || Empathy.empathies[:default]
-    empathy = current_user.empathies.create(post_id: post.id, empathy: empathy_value)
+    empathy = current_user.empathies.find_or_create_by(post_id: post.id) # Use find_or_create_by to ensure uniqueness
+    empathy.update(empathy: empathy_value)
     redirect_to request.referer
   end
 
   def destroy
     post= Post.find(params[:post_id])
     empathy = current_user.empathies.find_by(post_id: post.id)
-    empathy.destroy
+    empathy.destroy if empathy
     redirect_to request.referer
   end
 end
